+++ src/main/java/com/Chomusuke/gui/pane/TransactionPane.java
@@ -21,9 +21,9 @@ import com.chomusuke.Accountable.SceneID;
 import com.chomusuke.gui.element.DateSelector;
 import com.chomusuke.gui.element.PlusButton;
 import com.chomusuke.gui.element.SquareButton;
-import com.chomusuke.gui.element.Tile.TransactionTile;
-import com.chomusuke.gui.popups.AddFileScreen;
-import com.chomusuke.gui.popups.AddTransactionScreen;
+import com.chomusuke.gui.element.tile.TransactionTile;
+import com.chomusuke.gui.popup.AddFileScreen;
+import com.chomusuke.gui.popup.AddTransactionScreen;
 import com.chomusuke.logic.Account;
 import com.chomusuke.logic.Storage;
 import com.chomusuke.logic.Transaction;
