+++ src/main/java/com/Chomusuke/gui/popup/AddAccountScreen.java
@@ -15,7 +15,7 @@
     along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 
-package com.chomusuke.gui.popups;
+package com.chomusuke.gui.popup;
 
 import com.chomusuke.logic.Account;
 import com.chomusuke.logic.Storage;
