+++ src/main/java/com/Chomusuke/gui/popup/PopUp.java
@@ -15,7 +15,7 @@
     along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 
-package com.chomusuke.gui.popups;
+package com.chomusuke.gui.popup;
 
 import javafx.event.ActionEvent;
 import javafx.event.EventHandler;
